# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /videos:
    get:
      operationId: listVideos
      tags: [videos]
      summary: List all videos
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, processing, completed, failed, cancelled]
      responses:
        "200":
          description: List of videos
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Video"
  /videos/{id}:
    get:
      operationId: getVideo
      tags: [videos]
      summary: Get video details with frames and transcriptions
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Video details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VideoDetail"
        "404":
          description: Video not found
  /videos/{id}/tags:
    patch:
      operationId: updateVideoTags
      tags: [videos]
      summary: Update a video's tags
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateVideoTagsBody"
      responses:
        "200":
          description: Tags updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
        "404":
          description: Video not found
  /videos/{id}/process:
    post:
      operationId: processVideo
      tags: [videos]
      summary: Trigger video processing
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Processing started
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProcessingResponse"
        "404":
          description: Video not found
  /videos/{id}/cancel:
    post:
      operationId: cancelVideo
      tags: [videos]
      summary: Cancel video processing
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Video cancelled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProcessingResponse"
        "400":
          description: Video cannot be cancelled
        "404":
          description: Video not found
  /videos/sync:
    post:
      operationId: syncVideos
      tags: [drive]
      summary: Sync videos from a Google Drive folder
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SyncRequest"
      responses:
        "200":
          description: Sync initiated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SyncResponse"
...
components:
  schemas:
...
    Video:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        driveFileId:
          type: string
        driveFolderId:
          type: ["string", "null"]
        mimeType:
          type: string
        fileSize:
          type: ["integer", "null"]
        duration:
          type: ["integer", "null"]
        localPath:
          type: ["string", "null"]
        thumbnailPath:
          type: ["string", "null"]
        status:
          type: string
          enum: [pending, processing, completed, failed, cancelled]
        processingError:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - title
        - driveFileId
        - mimeType
        - status
        - createdAt
        - updatedAt
    VideoDetail:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        driveFileId:
          type: string
        driveFolderId:
          type: ["string", "null"]
        mimeType:
          type: string
        fileSize:
          type: ["integer", "null"]
        duration:
          type: ["integer", "null"]
        localPath:
          type: ["string", "null"]
        thumbnailPath:
          type: ["string", "null"]
        status:
          type: string
        processingError:
          type: ["string", "null"]
        tags:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        frames:
          type: array
          items:
            $ref: "#/components/schemas/FrameItem"
        transcriptions:
          type: array
          items:
            $ref: "#/components/schemas/TranscriptionItem"
      required:
        - id
        - title
        - driveFileId
        - mimeType
        - status
        - createdAt
        - updatedAt
        - frames
        - transcriptions
    FrameItem:
      type: object
      properties:
        id:
          type: integer
        videoId:
          type: integer
        timestampSec:
          type: number
        imagePath:
          type: string
        description:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - videoId
        - timestampSec
        - imagePath
        - createdAt
    TranscriptionItem:
      type: object
      properties:
        id:
          type: integer
        videoId:
          type: integer
        startSec:
          type: number
        endSec:
          type: number
        content:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - videoId
        - startSec
        - endSec
        - content
        - createdAt
    SyncRequest:
      type: object
      properties:
        folderId:
          type: string
      required:
        - folderId
    SyncResponse:
      type: object
      properties:
        syncedCount:
          type: integer
        videos:
          type: array
          items:
            $ref: "#/components/schemas/Video"
      required:
        - syncedCount
        - videos
...
    ProcessingResponse:
      type: object
      properties:
        message:
          type: string
        videoId:
          type: integer
        status:
          type: string
      required:
        - message
        - videoId
        - status
...
    UpdateVideoTagsBody:
      type: object
      properties:
        tags:
          type: string
      required:
        - tags
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
      required:
        - success
...