# lib/api-zod/src/generated/types/videoDetail.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { FrameItem } from "./frameItem";
import type { TranscriptionItem } from "./transcriptionItem";

export interface VideoDetail {
  id: number;
  title: string;
  driveFileId: string;
  /** @nullable */
  driveFolderId?: string | null;
  mimeType: string;
  /** @nullable */
  fileSize?: number | null;
  /** @nullable */
  duration?: number | null;
  /** @nullable */
  localPath?: string | null;
  /** @nullable */
  thumbnailPath?: string | null;
  status: string;
  /** @nullable */
  processingError?: string | null;
  /** @nullable */
  tags?: string | null;
  createdAt: Date;
  updatedAt: Date;
  frames: FrameItem[];
  transcriptions: TranscriptionItem[];
}